+++ NEWS.md
@@ -2,6 +2,8 @@ Changes in 0.4.1
 ================
 
 -   lexpand bug fixed (\#120): observations were dropped if their entry by age was smaller than the smallest age value, though entry at exit is correct and used now.
+-   `sir` rewrite. New more consistent output, updates on plotting and minor changes in arguments. Introduce very simple `coef()` and `confint()` methods for sir class.
+-   new functions in sir family: `sir_ag()`, `sir_lex()` and `sir_exp()` for extracting SMRs from `aggre` and `Lexis` objects.
 
 Changes in 0.4.0
 ================
@@ -50,9 +52,6 @@ sr$status <- factor(sr$status, levels = 0:2,
 library(Epi)
 #> 
 #> Attaching package: 'Epi'
-#> The following object is masked from 'package:popEpi':
-#> 
-#>     [.Lexis
 #> The following object is masked from 'package:base':
 #> 
 #>     merge.data.frame